pyproject.toml
[tool.poetry]
name = "dataliteracy"
version = "0.1.0"
description = ""
authors = ["Robin Uhrich <robin.uhrich@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.26.2"
requests = "^2.31.0"
schedule = "^1.2.1"
pandas = "^2.1.4"
pyaml = "^23.9.7"
wget = "^3.2"
shapely = "^2.0.2"
geopandas = "^0.14.1"
matplotlib = "^3.8.2"
tqdm = "^4.66.1"
urllib3 = "<2"
scikit-learn = "^1.3.2"

[tool.poetry.group.dev.dependencies]
pyargwriter = {git = "https://github.com/RobinU434/PyArgWriter"}
jupyter = "^1.0.0"
ruff = "^0.1.8"
pytest = "^7.4.3"
sqlalchemy = "^2.0.25"

[tool.poetry.group.crawler]
optional = true

[tool.poetry.group.crawler.dependencies]
mariadb = "^1.1.8"
sqlalchemy = "^2.0.23"


[tool.poetry.group.sam_explore.dependencies]
polars = "^0.20.1"
geopy = "^2.4.1"


[tool.poetry.group.playwright.dependencies]
pytest-playwright = "^0.4.3"
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
log_cli = true